moved view manager to a dedicated location
